+++ crates/apollo-compiler/tests/validation/types.rs
@@ -1,8 +1,10 @@
 //! Ported from graphql-js, 2023-11-16
 //! https://github.com/graphql/graphql-js/blob/0b7590f0a2b65e6210da2e49be0d8e6c27781af2/src/validation/__tests__/ValuesOfCorrectTypeRule-test.ts
+use apollo_compiler::schema::SchemaBuilder;
 use apollo_compiler::validation::Valid;
 use apollo_compiler::ExecutableDocument;
 use apollo_compiler::Schema;
+use expect_test::expect;
 use expect_test::Expect;
 use std::sync::OnceLock;
 use unindent::unindent;
@@ -2015,4 +2017,35 @@ mod variable_default_values {
             "#]],
         );
     }
+}
+
+#[test]
+fn handles_built_in_type_redefinition() {
+    let schema = r#"
+scalar String
+
+type Query {
+  foo: String
+}
+"#;
+
+    let errors = Schema::parse_and_validate(schema, "schema.graphql")
+        .expect_err("invalid schema")
+        .errors;
+    let expected = expect![[r#"
+        Error: built-in scalar definitions must be omitted
+           ╭─[ schema.graphql:2:1 ]
+           │
+         2 │ scalar String
+           │ ──────┬──────  
+           │       ╰──────── remove this scalar definition
+        ───╯
+    "#]];
+    expected.assert_eq(&errors.to_string());
+
+    let builder = SchemaBuilder::new().ignore_builtin_redefinitions();
+    let _ = builder
+        .parse(schema, "schema.graphql")
+        .build()
+        .expect("schema parsed successfully");
 }